+++ yeonjae22/ElevenStreetSample/app/src/main/AndroidManifest.xml
@@ -10,7 +10,8 @@
             android:label="@string/app_name"
             android:roundIcon="@mipmap/ic_launcher_round"
             android:supportsRtl="true"
-            android:theme="@style/AppTheme">
+            android:theme="@style/AppTheme"
+            android:name=".util.App">
         <activity android:name=".ProductActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN"/>
