+++ yeonjae22/ElevenStreetSample/app/src/main/java/com/exam/elevenstreet/util/App.kt
@@ -1,0 +1,18 @@
+package com.exam.elevenstreet.util
+
+import android.app.Application
+import android.content.Context
+
+class App : Application() {
+    companion object {
+        lateinit var instance: App
+            private set
+    }
+
+    override fun onCreate() {
+        super.onCreate()
+        instance = this
+    }
+
+    fun context(): Context = applicationContext
+}
