+++ yeonjae22/ElevenStreetSample/app/src/main/java/com/exam/elevenstreet/view/product/ProductActivity.kt
@@ -7,26 +7,18 @@ import androidx.recyclerview.widget.LinearLayoutManager
 import com.ElevenStreetApi
 import com.exam.RetrofitInstance
 import com.exam.elevenstreet.data.ProductLocalDataSource
+import com.exam.elevenstreet.data.ProductRemoteDataSource
 import com.example.elevenstreet.ProductResponse
-import com.example.elevenstreet.ProductXmlPullParserHandler
 import kotlinx.android.synthetic.main.activity_main.*
-import java.net.URL
-
 
 class ProductActivity : AppCompatActivity() {
 
     private var elevenStreetApi: ElevenStreetApi? = null
     val adapter = ProductAdapter()
 
-    interface CallBack {
-        fun onSuccess(productList1: List<ProductResponse>)
-        fun onFailure(message: String)
-    }
-
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
-
         elevenStreetApi =
             RetrofitInstance.getInstance<ElevenStreetApi>("https://openapi.11st.co.kr/openapi/")
 
@@ -34,36 +26,20 @@ class ProductActivity : AppCompatActivity() {
         recycler_view.setLayoutManager(manager)
         recycler_view.setHasFixedSize(true)
 
-        val productList = ProductLocalDataSource().getProductList(applicationContext)
-        //val adapter = ProductAdapter()
+        val productList = ProductLocalDataSource().getProductList()
         recycler_view.adapter = adapter
         adapter.addData(productList)
 
         btn_search.setOnClickListener {
-            ProductTask().execute()
+            ProductTask().execute("${edt_search.text}")
         }
 
     }
 
-    fun getSearchByKeyword(
-    ): List<ProductResponse> {
-        val call = elevenStreetApi?.getProductList(
-            getString(R.string.eleven_street_API_KEY),
-            API_CODE,
-            "${edt_search.text}",
-            1
-        )
-        val url: String? = "${call?.request()?.url()}"
-        val targetURL = URL(url)
-        val inputStream = targetURL.openStream()
-        val productList = ProductXmlPullParserHandler().parse(inputStream)
-        return productList
-    }
-
-    internal inner class ProductTask : AsyncTask<String, Void, List<ProductResponse>>() {
+    inner class ProductTask : AsyncTask<String, Void, List<ProductResponse>>() {
 
         override fun doInBackground(vararg keyWords: String): List<ProductResponse>? {
-            val productList1 = getSearchByKeyword()
+            val productList1 = ProductRemoteDataSource().getSearchByKeyword(keyWords[0])
             return productList1
         }
